requirements.txt
alembic==1.4.3
Babel==2.9.0
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.6
python-dateutil==2.6.0
python-editor==1.0.4
pytz==2020.5
six==1.15.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
WTForms==2.3.3
